Removed cameraOps and created two methods, closeCamera & startCamera that use basic logic to detect if Motion is currently running. If it is, a button to close it will appear. If not, a button to open it will appear.
Home page now refreshes the snapshot immediately and then at an interval, AND determines if Motion is already running or not, and displays the appropriate button.

+++ app/controllers/rubytest_controller.rb
@@ -3,7 +3,7 @@ class RubytestController < ApplicationController
   before_filter :authenticate_user
   
   def index
-    
+        
   end
   
   def left
@@ -47,30 +47,69 @@ class RubytestController < ApplicationController
     render :partial => "rubytest/refreshSnapshot"
   end
   
-  def cameraOps
-    @status = "No change."
+  def closeCamera
+    # Initial check for motion.pid file
+    pid_exists = File.exist?("/var/www/test/pid/motion.pid")
     
-    destroy_process = system("kill $(pidof motion)")
-
-    if destroy_process
-
-        @status = "Motion destroyed."
-        destroy_pid_file = system("rm -f $(find /var/www/test -name 'motion.pid')")
+    if pid_exists
+      # If PID file exists, destroy process first so that it has a graceful quit
+      destroy_process = system("kill $(pidof motion)")
+      
+      # Check again if the motion.pid exists
+      # If it does, then process wasn't killed gracefully
+      pid_still_exists = File.exist?("/var/www/test/pid/motion.pid")
+      
+      if pid_still_exists
+        # Remove the file if it exists
+        remove_pid_forcefully = File.delete("/var/www/test/pid/motion.pid")
         
-        if destroy_pid_file
-          @status = "Motion process destroyed; motion.pid removed forcefully."
-          @button_label = "Turn on camera"
+        if remove_pid_forcefully
+          @status = "PID file removed forcefully."
+          @button_label = "Start Motion"  
         end
-
-    else
-      start = system("/usr/local/bin/motion -m")
-      @status = "Static Camera Started."
-      @button_label = "Turn off camera."
+        
+      else
+        
+        @status = "Motion gracefully destroyed."
+        @button_label = "Start Motion"
+        
+      end
       
     end
     
     respond_to do | format |
-        format.html   { redirect_to index_path }
+        format.html
+        format.js
+    end
+    
+  end
+  
+  def startCamera
+    # Initial check for motion.pid file
+    pid_exists = File.exist?("/var/www/test/pid/motion.pid")
+    
+    # Ensure that motion is NOT running
+    if !pid_exists
+      
+      start = system("/usr/local/bin/motion -m")
+      
+      @pid_now_exists = File.exist?("/var/www/test/pid/motion.pid")
+      
+      if @pid_now_exists
+        @status = "Static Camera Started."
+        @button_label = "Close Motion."
+      else
+        @status = "Motion failed to start."
+        @button_label = "Start Motion"
+      end
+      
+    else
+      @status = "Motion is already running."
+      @button_label = "Close Motion"
+    end
+    
+    respond_to do | format |
+        format.html
         format.js
     end
